Remove analytics responsibilities from Pushhandler and move them into separate class
